--- Assets/Prefabs/Bot.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 487f0c203d4e587499141d56e06a867f
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
