--- Assets/Scenes/Bots.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3d3b215d0b631cb4790c4c2fb04ee19e
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
