--- Assets/Scripts/Bot.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6dbaa59dc029243588bddd79f52088da
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
