--- Assets/Scripts/Bot/Bot.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5022502e6c076ab4e90e8aa17cde4c44
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
