--- Assets/Scripts/Bot/BotBrain.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fb19ed6abf2ff42de8975e8aedcaa974
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
