--- Assets/Scripts/Test.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using System.Linq;
-
-public class Test : MonoBehaviour
-{
-
-}
