--- Assets/Scripts/Test.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 75ad4728e167e464696a4b5dbd72ab27
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
