--- Assets/Scripts/UI/NetworkView.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e3dde2991fd3525409f44bb33338d06d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
