--- Assets/Terrain/Bot Terrain.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 95ecc7aa9dfcff442ab398e33c6b8a2c
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 15600000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
