--- Logs/shadercompiler-UnityShaderCompiler.exe0.log
@@ -1,3 +1,0 @@
-Base path: 'C:/Program Files/2021.1.1f1/Editor/Data', plugins path 'C:/Program Files/2021.1.1f1/Editor/Data/PlaybackEngines'
-Cmd: initializeCompiler
-
